server/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.0.1
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
